moves .../algebra/ops/extexprs/ExtensionFunction.java -> .../algebra/ops/extfcts/ExtensionFunction.java
