+++ hefquin-rml/src/main/java/se/liu/ida/hefquin/mappings/algebra/ops/extfcts/ExtnFct_ToIRI.java
@@ -8,7 +8,6 @@ import org.apache.jena.graph.Node;
 import org.apache.jena.graph.NodeFactory;
 
 import se.liu.ida.hefquin.mappings.algebra.MappingRelation;
-import se.liu.ida.hefquin.mappings.algebra.ops.extexprs.ExtensionFunction;
 
 public class ExtnFct_ToIRI implements ExtensionFunction
 {
