+++ hefquin-rml/src/main/java/se/liu/ida/hefquin/mappings/algebra/ops/extfcts/ExtnFct_ToLiteral.java
@@ -6,7 +6,6 @@ import org.apache.jena.graph.Node;
 import org.apache.jena.graph.NodeFactory;
 
 import se.liu.ida.hefquin.mappings.algebra.MappingRelation;
-import se.liu.ida.hefquin.mappings.algebra.ops.extexprs.ExtensionFunction;
 
 /**
  * Given a string literal and a datatype IRI, returns a literal whose
